Fix interlaced resize and add some benchmarks
Interlaced resize was creating invalid slices at the last iteration
which will panic, even if the slice is never accessed.
Instead, we maintain an index and never access outside the slice.
Furthermore, with interlaced resize, the last line is width pixels,
not pitch pixels, which will overread/overwrite on the second field.